Move CLI output to own package, refactor, update check.go
